pull-socket: Extract dummy signature handling
